Add Orkut test.

+++ include/graph.h
@@ -23,17 +23,17 @@ public:
     int intersection_size(int v1,int v2);
 
     //single thread triangle counting
-    int triangle_counting();
+    long long triangle_counting();
     
     //multi thread triangle counting
-    int triangle_counting_mt(int thread_count);
+    long long triangle_counting_mt(int thread_count);
 
     //general pattern matching algorithm with multi thread
-    int pattern_matching(const Schedule& schedule, int thread_count);
+    long long pattern_matching(const Schedule& schedule, int thread_count);
 private:
-    void tc_mt(int * global_ans);
+    void tc_mt(long long * global_ans);
 
     void get_edge_index(int v, int& l, int& r) const;
 
-    void pattern_matching_func(const Schedule& schedule, VertexSet* vertex_set, VertexSet& subtraction_set, int& local_ans, int depth);
+    void pattern_matching_func(const Schedule& schedule, VertexSet* vertex_set, VertexSet& subtraction_set, long long& local_ans, int depth);
 };
